move plugin.py so name doesn't conflict with c module
